Refactored DatasetBundler as base class
